Convert EC2 utilities to Streamline.js.
Converted all of the utilities that use `node-ec2` to Streamline.js.

It is for this application that I submitted the patch to have Streamline.js run
an interpreter, with registered file extensions, so that it could be used for
system administration, where the script may be installed in a directory that is
not writable by the user who invokes the script.

Streamline.js was generating temporary files, then requiring the temporary
files. This was causing problems for the conversion to Streamline.js.

After conversion, the programs are much easier to read. Probably easier to
maintain. Time will tell.

I've not had the opportunity to run every last script as part of real system
administration tasks. There may be bugs in this commit. Sorry.