Deploy Oregon.
Create a Makefile that will build Oregon. The Makefile simply touches files to
indicate that a step in the deployment has completed. Not at all complete. I'm
imagining a daily launch of an entire region, although since starting this
patch, I believe it's more likely that I'll use AWS California for testing,
Oregon for deployment, and Virginia for secondary DNS.