Launch with AWS security group.
When launching an image, `bin/instance/launch` will now specify an AWS security
group. The security group is specified by name, using the name of the instance
type; `balance`, `user`, `data`, `image`, etc.

For the instance to launch successfully, a security group with the instance type
name must be defined for the AWS region.