Simplify `unbound.conf`.
The previous `unbound.conf` was inspired by the one that Fedora installs. That
`unbound.conf` specifies values for some of the more obscure settings. Some of
those values are the default values, so those entries are not needed. Some of
those values are non-default values which cause Unbound so much grief on
VirtualBox, all requests timeout.

The `harden-referral-path` switch causes Unbound to "[enforce] DNSSEC validation
on nameserver NS sets and the nameserver addresses that are encountered on the
referral path to the answer." Non-standard and a lot of work.

Disabled `use-caps-for-id` because some servers don't do dns-0x20 property.

The default for `val-log-level` is off, but it is interesting, recording
validation failures regardless of the log level.

Otherwise, the specified values of the settings removed are the defaults.