Change directory to Puppy root.
Instead of setting `PUPPYIFY_PATH`, I change directory to the root directory of
the janitor. Then I can call all programs relative to `PUPPIFY_PATH`.

This is going to make it difficult to ever install `puppify` in a UNIX file
hierarchy, but it is going to make the bash programs easier to read. When I have
to use variables to gather up path locations, my code starts to look like it is
variables only. You end up with ugliness such as this.

```
$PUPPY_PATH $PUPPY_EXEC $PUPPY_OPTIONS > $PUPPY_OUTPUT
```

Although, it might be possible to wrap up the invocations in a function, and
that function can do the right thing, giving me an opportunity to reorganize the
project directory and file name layout later.