The `PUPPY_LIBRARY` experiment.
Even better than using relative paths, I've not created a function called
`puppy_exec` that will accept a relative path, then it will invoke the program
at the relative path. This gives me the syntax I want. If I change the project
layout later, I don't have to audit my code to alter all invocations. If I
decide to keep all of the executables in the same `bin` directory, then I can
simply convert the slash delimited executable name into a hyphen delimited
executable name.

One of the bits that I was agonizing over was how to get a new program started
without a long and involved preamble to find a common set of functions. I had
the bring of of creating a `PUPPY_LIBRARY` environment variable in
`bin/puppify`, exporting to it can be evaluated. I'm considering the security
implications of this, but it looks as though it is no more or less secure than
anything the user can already do.

This terse preamble means you can't simply put these things in the path, since
they will not execute with the environment variable set. It makes the
executables dependent on the wrapper, so we may end up saying they are library
executables, and remove them from the path.