Converting install scripts, split up pretty good.
Step one to creating a `puppify` program that will use the pivot program style
of `git`, `yum`, or `svn`, is to create that pivot program, and to break up our
monolithic program into library programs.

I've created `lib/puppy/bin` to house the library programs. They are not to be
run directly, but through the `bin/puppify` pivot program.

Breaking up programs in this way is going go to work really well with SELinux.