Letting go of obligations.
Not commenting a dozen small commits was eating at me, so I squashed them. This
is the bulk of the rewrite of `bin/puppify`. New is a `Makefile` generator, that
creates a `Makefile` that keeps track of the progress of the build.

It simply touches status files in a `./made` directory, it doesn't actually
compare source files to their destinations. We don't need all that. It's easier
to express the build as a `Makefile`.