Converted `checkout` to a `puppify` sub-program.
Converted `checkout` to a `puppyify` sub-program. Starting to reach a
point in this process where it's just fiddly. Not sure that after this
change is made, that all of these programs will survive. I'm going to
start building some servers with the `bin/puppify` program, to get
feedback on its design.